+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Add verify encoding script
 - Rename folders
 - Move Windows Server 2022 image to separate folder
+- Update packer images folder structure
+- Add packer init script
 
 ## v1.0.0 - In Progress